pyproject.toml
[tool.poetry]
name = "dailygroove"
version = "0.1.0"
description = "A music guessing game"
authors = ["lekjos <leif@leifkjos.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
crispy-bootstrap4 = "*"
django = "^3.2"
django-anymail = "^10.0"
django-crispy-forms = "*"
django-extensions = "^3.2.3"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
django-timezone-field = "^5.1"

[tool.poetry.group.dev.dependencies]
black= "^23.3.0"
coverage = "*"
debugpy = "*"
django-debug-toolbar = "*"
factory-boy = "^3.2.1"
faker = "^19.1.0"
isort = "^5.12.0"
mixer = "*"
pre-commit = "^3.3.3"
pylint = "^2.17.4"
pylint-django = "^2.5.3"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
py_version = 310
src_paths =  ["src"]
include_trailing_comma = true
profile = "black"
default_section = "THIRDPARTY"
known_first_party = ["src"]
known_django = ["django"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]